consensus_types

Module block

Structs

BlockDigest

Digest of a VerifiedBlock or verified SignedBlock, which covers the Block and its signature.

BlockRef

BlockRef uniquely identifies a VerifiedBlock via digest. It also contains the slot info (round and author) so it can be used in logic such as aggregating stakes for a round.

Constants

NUM_RESERVED_TRANSACTION_INDICES

The number of reserved transaction indices.

PING_TRANSACTION_INDEX

The index of the ping transaction. This is a special type of transaction that is used in order to ping consensus and perform ping requests. We reserve the max index for this transaction.

Type Aliases

BlockTimestampMs

Block proposal timestamp in milliseconds.

Round

Round number of a block.

TransactionIndex

Index of a transaction in a block.