Function consensus_types::block::block_test_utils::new_proposal

#[must_use = "strategies do nothing unless used"]
pub fn new_proposal(
    _ancestor_id: HashValue, 
    round_strategy: impl Strategy<Value = Round>, 
    signer_strategy: impl Strategy<Value = ValidatorSigner>, 
    parent_qc: QuorumCert
) -> impl Strategy<Value = Block<Vec<usize>>>

This strategy is a swiss-army tool to produce a low-level block dependent on signer, round, parent and ancestor_id. Note that the quorum certificate carried by this block is still placeholder: one will have to generate it later on when adding to the tree.