Module consensus_types::block_retrieval

Structs

BlockRetrievalRequest
RPC to get a chain of block of the given length starting from the given block id.

BlockRetrievalResponse
Carries the returned blocks and the retrieval status.

Enums

BlockRetrievalStatus