Struct consensus_types::timeout_2chain::TwoChainTimeout

pub struct TwoChainTimeout { /* fields omitted */ }

This structure contains all the information necessary to construct a signature on the equivalent of a DiemBFT v4 timeout message.

Implementations

impl TwoChainTimeout

pub fn new(epoch: u64, round: Round, quorum_cert: QuorumCert) -> Self

pub fn epoch(&self) -> u64

pub fn round(&self) -> Round

pub fn hqc_round(&self) -> Round

pub fn quorum_cert(&self) -> &QuorumCert

pub fn sign(&self, signer: &ValidatorSigner) -> Ed25519Signature

pub fn signing_format(&self) -> TimeoutSigningRepr

pub fn verify(&self, validators: &ValidatorVerifier) -> Result<()>