Module consensus_types::block::block_test_utils

Functions

block_forest_and_its_keys
This returns keys and a block forest created from them

certificate_for_genesis

fake_id
This produces a block with an invalid id (and therefore signature) given a valid block

gen_test_certificate

genesis_strategy
This produces the genesis block

get_current_timestamp

leaf_strategy
Offers the genesis block.

new_proposal
This strategy is a swiss-army tool to produce a low-level block dependent on signer, round, parent and ancestor_id. Note that the quorum certificate carried by this block is still placeholder: one will have to generate it later on when adding to the tree.

placeholder_certificate_for_block

placeholder_ledger_info

some_round
This produces a round that is often higher than the parent, but not too high

unmoored_block
This produces an unmoored block, with arbitrary parent & QC ancestor